+++ packages/circuit-ui/components/ColorInput/ColorInput.stories.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019, SumUp Ltd.
+ * Copyright 2024, SumUp Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,6 +13,22 @@
  * limitations under the License.
  */
 
-export { Input, classes } from './Input.js';
+import { ColorInput, type ColorInputProps } from './ColorInput.js';
 
-export type { InputProps, BaseInputProps, InputElement } from './Input.js';
+export default {
+  title: 'Forms/ColorInput',
+  component: ColorInput,
+};
+
+const baseArgs = {
+  label: 'Color',
+  pickerLabel: 'Pick color',
+  placeholder: '#99ffbb',
+  defaultValue: '#99ffbb',
+};
+
+export const Base = (args: ColorInputProps) => (
+  <ColorInput {...args} style={{ maxWidth: '250px' }} />
+);
+
+Base.args = baseArgs;