+++ packages/circuit-ui/components/PhoneNumberInput/PhoneNumberInput.spec.tsx
@@ -24,8 +24,9 @@ import {
 } from './PhoneNumberInput.js';
 
 const countryCodeMap: { [key: string]: string } = {
-  '+1': 'US',
-  '+49': 'DE',
+  CA: '+1',
+  US: '+1',
+  DE: '+49',
 };
 
 const defaultProps = {
@@ -34,8 +35,8 @@ const defaultProps = {
     label: 'Country code',
     defaultValue: '+1',
     options: Object.keys(countryCodeMap).map((key) => ({
-      country: countryCodeMap[key],
-      code: key,
+      country: key,
+      code: countryCodeMap[key],
     })),
   },
   subscriberNumber: {
@@ -84,7 +85,7 @@ describe('PhoneNumberInput', () => {
     };
     render(<PhoneNumberInput {...props} />);
     const select = screen.getByRole('combobox');
-    await userEvent.selectOptions(select, '+49');
+    await userEvent.selectOptions(select, 'DE');
     expect(onChange).toHaveBeenCalledOnce();
   });
 
@@ -99,7 +100,7 @@ describe('PhoneNumberInput', () => {
     };
     render(<PhoneNumberInput {...props} />);
     const select = screen.getByRole('combobox');
-    await userEvent.selectOptions(select, '+49');
+    await userEvent.selectOptions(select, 'DE');
     expect(onChange).toHaveBeenCalledOnce();
   });