+++ packages/circuit-ui/components/PhoneNumberInput/PhoneNumberInput.tsx
@@ -240,8 +240,20 @@ export const PhoneNumberInput = forwardRef<
       ) {
         return;
       }
+
+      const selectedCountry = countryCodeRef?.current?.value;
+      if (!selectedCountry) {
+        return;
+      }
+      const code = countryCode.options.find(
+        ({ country }) => country === selectedCountry,
+      )?.code;
+
+      if (!code) {
+        return;
+      }
       const phoneNumber = normalizePhoneNumber(
-        countryCodeRef.current.value,
+        code,
         subscriberNumberRef.current.value,
       );
       onChange(phoneNumber);
@@ -273,21 +285,20 @@ export const PhoneNumberInput = forwardRef<
         return;
       }
 
-      const pastedCountryCode = countryCode.options
-        .map((option) => option.code)
+      const pastedOption = countryCode.options
         // Match longer, more specific country codes first
-        .sort((a, b) => b.length - a.length)
-        .find((code) => pastedText.startsWith(code));
+        .sort((a, b) => b.code.length - a.code.length)
+        .find(({ code }) => pastedText.startsWith(code));
 
-      if (!pastedCountryCode) {
+      if (!pastedOption) {
         return;
       }
 
       event.preventDefault();
 
-      const pastedSubscriberNumber = pastedText.split(pastedCountryCode)[1];
+      const pastedSubscriberNumber = pastedText.split(pastedOption.code)[1];
 
-      countryCodeRef.current.value = pastedCountryCode;
+      countryCodeRef.current.value = pastedOption.country;
 
       // React overwrites the input.value setter. In order to be able to trigger
       // a 'change' event on the input, we need to use the native setter.